feat(backups): Introduce BackupPlanManager for managing backup plans and jobs
- Added BackupPlanManager class to handle creation and management of backup plans and jobs.
- Implemented methods for creating Backrest plans and Minio bucket backup jobs.
- Integrated with OPClient to retrieve backup server details and Volsync password.
- Updated various stacks to utilize the new BackupPlanManager for backup job creation.
- Removed obsolete ProxmoxBackupServer and backups.ts files.
- Updated tsconfig files to include new BackupPlanManager module.
- Added tags to ProxmoxBackupServerLxc instances to differentiate between backup sources and destinations.
