--- stacks/backups/backups.ts
@@ -1,141 +1,0 @@
-import { DockgeLxc } from "../../components/DockgeLxc.ts";
-import * as pulumi from "@pulumi/pulumi";
-import { ProxmoxHost } from "../../components/ProxmoxHost.ts";
-import { CategoryEnum, OnePasswordItem as OPI, TypeEnum } from "@dynamic/1password/OnePasswordItem.ts";
-import { GlobalResources } from "@components/globals.ts";
-import * as minio from "@pulumi/minio";
-import { BackupPlanManager } from "../../components/jobs.ts";
-
-export function createBackupJobs({ globals }: { globals: GlobalResources }) {
-  // const celestiaBackupManager = new BackupPlanManager("celestia-backup-plan-manager", {
-  //   globals: globals,
-  //   source: source,
-  //   localBackup: source,
-  //   remoteBackup: destination,
-  // });
-
-  // celestiaBackupManager.createBackrestPlan("immich", {
-  //   title: "Immich",
-  //   paths: ["/spike/data/immich/"],
-  //   repository: "immich",
-  //   planConfig: {
-  //     excludes: ["/spike/data/immich/backups", "/spike/data/immich/encoded-video"],
-  //   },
-  // });
-
-  // celestiaBackupManager.createBackrestPlan("pgdump", {
-  //   title: "Postgres Dumps",
-  //   paths: ["/spike/data/pgdump/"],
-  //   repository: "pgdump",
-  // });
-
-  createMinioBucketBackupJob({
-    title: "Home Operations",
-    bucket: "home-operations",
-    // backblazeSecret: "Backblaze home-operations",
-    // source: source,
-    // destination: destination,
-    globals,
-  });
-  createMinioBucketBackupJob({
-    title: "Stargate Command Postgres",
-    bucket: "stargate-command-db",
-    restoreBucket: "stargate-command-db-restore",
-    // backblazeSecret: "Backblaze S3 Stargate Command Database",
-    // source: source,
-    // destination: destination,
-    globals,
-  });
-  createMinioBucketBackupJob({
-    title: "Equestria Postgres",
-    bucket: "equestria-db",
-    restoreBucket: "equestria-db-restore",
-    // backblazeSecret: "Backblaze S3 Equestria Database",
-    // source: source,
-    // destination: destination,
-    globals,
-  });
-
-  // Backup Thanos bucket to Celestia and Luna
-  // createMinioBucketBackupJob({ title: "Thanos Storage", bucket: thanosStorage.bucket, source: source, destination: destination, globals });
-
-  // await alphaSiteBackupManager.updateBackrestConfig();
-  // TODO: Handle luna backrest config
-  // celestiaBackupManager.updateBackrestConfig();
-}
-
-function createMinioBucketBackupJob({
-  // title,
-  bucket,
-  // backblazeSecret,
-  // source,
-  // destination,
-  globals,
-  restoreBucket: restoreBucket,
-}: {
-  // source: DockgeLxc;
-  // destination: DockgeLxc;
-  title: pulumi.Input<string>;
-  bucket: pulumi.Input<string>;
-  globals: GlobalResources;
-  backblazeSecret?: pulumi.Input<string>;
-  restoreBucket?: pulumi.Input<string>;
-}) {
-  // source.createBackupJob({
-  //   name: pulumi.interpolate`Backup ${title}`,
-  //   schedule: "0 10 * * *",
-  //   sourceType: "local",
-  //   source: pulumi.interpolate`/spike/data/minio/${bucket}/`,
-  //   destinationType: "local",
-  //   destination: pulumi.interpolate`/data/backup/spike/${bucket}/`,
-  // });
-
-  // if (backblazeSecret) {
-  //   source.createBackupJob({
-  //     name: pulumi.interpolate`Replicate ${title} to B2`,
-  //     schedule: "0 8 */4 * *",
-  //     sourceType: "local",
-  //     source: pulumi.interpolate`/spike/data/minio/${bucket}/`,
-  //     destinationType: "b2",
-  //     destination: pulumi.interpolate`/`,
-  //     destinationSecret: backblazeSecret,
-  //   });
-  // }
-
-  // destination.createBackupJob({
-  //   name: pulumi.interpolate`Replicate ${title}`,
-  //   schedule: "0 3 * * *",
-  //   sourceType: "sftp",
-  //   source: pulumi.interpolate`${source.tailscaleHostname}/spike/${bucket}/`,
-  //   destinationType: "local",
-  //   destination: pulumi.interpolate`/data/backup/spike/${bucket}/`,
-  // });
-
-  if (restoreBucket) {
-    pulumi.all([bucket, restoreBucket]).apply(([bucket, restoreBucket]) => {
-      const minioBucket = new minio.S3Bucket(
-        `${restoreBucket}-minio-bucket`,
-        {
-          acl: "private",
-          bucket: pulumi.interpolate`${restoreBucket}`,
-        },
-        {
-          provider: globals.truenasMinioProvider,
-          protect: true,
-          retainOnDelete: true,
-          import: restoreBucket,
-        },
-      );
-      // source.createBackupJob({
-      //   name: pulumi.interpolate`Sync ${bucket} from ${restoreBucket}`,
-      //   schedule: "*/10 * * * *",
-      //   sourceType: "s3",
-      //   source: pulumi.interpolate`${bucket}/`,
-      //   sourceSecret: globals.truenasMinioCredential.title!,
-      //   destinationType: "s3",
-      //   destination: pulumi.interpolate`${minioBucket.bucket}/`,
-      //   destinationSecret: globals.truenasMinioCredential.title!,
-      // });
-    });
-  }
-}
