+++ stacks/unifi-network/index.ts
@@ -1,5 +1,5 @@
-import { GlobalResources } from "../../components/globals";
-import { createTailscaleAttDropFirewallRule } from "./tailscale-drop-firewall-rule";
+import { GlobalResources } from "../../components/globals.ts";
+import { createTailscaleAttDropFirewallRule } from "./tailscale-drop-firewall-rule.ts";
 import { assignTailscaleAcls } from "./acl-manager.ts";
 
 const globals = new GlobalResources("global-resources", {});
