+++ stacks/unifi-network/tsconfig.json
@@ -34,6 +34,7 @@
     "../../dynamic/**/*.ts",
     "../../types/**/*.ts",
     "../home/jobs.ts",
-    "../home/backup-jobs.ts"
+    "../home/backup-jobs.ts",
+    "../backups/BackupPlanManager.ts"
   ]
 }
