+++ tsconfig.json
@@ -30,5 +30,6 @@
     "components/**/*.ts",
     "components/**/*.mts",
     "types/**/*.ts",
+    "stacks/backups/BackupPlanManager.ts",
   ],
 }
